+++ plugins/framework/skills/framework-initialization/scripts/memory/lib/core/index.js
@@ -6,9 +6,11 @@
  * @license BSD-3-Clause
  */
 import EnvironmentManager from './environment.js';
+import GitHubAuth from './auth.js';
 import MemoryBuilderError from './error.js';
 
 export {
   EnvironmentManager,
+  GitHubAuth,
   MemoryBuilderError
 };